Error Handling

+++ src/main/java/com/sampa/springapi/controller/UserController.java
@@ -4,6 +4,9 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.core.Authentication;
@@ -19,6 +22,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.sampa.springapi.UserInfoUserDetails;
+import com.sampa.springapi.exceptions.ResponseMessage;
 import com.sampa.springapi.model.User;
 import com.sampa.springapi.repository.UserRepository;
 import com.sampa.springapi.service.JwtService;
@@ -37,24 +41,36 @@ public class UserController {
 	
 	@PostAuthorize("hasRole('ROLE_ADMIN')")
 	@GetMapping("/users")
-	public List<User> getUsers() {
-		return userService.getAllUsers();
+	public ResponseEntity<List<User>> getUsers() {
+		try {
+			List<User> users = userService.getAllUsers();
+			if(users.isEmpty()) {
+				return ResponseEntity.noContent().build();
+			}
+			return ResponseEntity.ok(users);
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+		}
 	}
 	
 	@PostAuthorize("hasRole('ROLE_STUDENT')")
 	@GetMapping("/users/{id}")
-	public User getUserById(@PathVariable Long id, HttpServletRequest request) {
-		String token = request.getHeader("Authorization").substring(7);
-		System.out.println("============================" + token + "===========================");
-		Integer userId = jwtService.extractUserId(token);
-		
-		System.out.println("============================" + userId + "===========================");
-	    if (!userId.equals(id.intValue())) {
-	        throw new AccessDeniedException("User not authorized to access this resource");
-	    }
-	    return userService.getUserByid(id);
+	public ResponseEntity<?> getUserById(@PathVariable Long id, HttpServletRequest request) {
+		try {
+			String token = request.getHeader("Authorization").substring(7);
+			Integer userId = jwtService.extractUserId(token);
+		    if (!userId.equals(id.intValue())) {
+		        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ResponseMessage("User not authorized to access this resource"));
+		    }
+		    User users = userService.getUserByid(id);
+		    return ResponseEntity.ok(users);
+		    
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+		}
+			
+		}
 		      
-	}
 	
 	@PostAuthorize("hasRole('ROLE_ADMIN')")
 	@PostMapping("/users")
@@ -64,8 +80,10 @@ public class UserController {
 	
 	@PostAuthorize("hasRole('ROLE_STUDENT')")
 	@PutMapping("/users/{id}")
-	public User updateUserById(@PathVariable Long id, @RequestBody User userDetails) {
+	public User updateUserById(@PathVariable Long id, @RequestBody User userDetails, HttpServletRequest request) {
 		User  user = userService.updateUserById(id, userDetails);
+		String token = request.getHeader("Authorization").substring(7);
+		Integer userId = jwtService.extractUserId(token);
 		if(userDetails.getUsername() != null) {
 			user.setUsername(userDetails.getUsername());
 		}
@@ -76,11 +94,20 @@ public class UserController {
 			user.setPassword(userDetails.getPassword());
 		}
 		
+		if(!userId.equals(id.intValue())) {
+			throw new AccessDeniedException("User not authorized to access this resource");
+		}
 		return userService.createUser(user);
 	}
 	
+	@PostAuthorize("hasRole('ROLE_STUDENT')")
 	@DeleteMapping("/users/{id}")
-	public String deleteUserById(@PathVariable Long id) {
+	public String deleteUserById(@PathVariable Long id, HttpServletRequest request) {
+		String token = request.getHeader("Authorization").substring(7);
+		Integer userId = jwtService.extractUserId(token);
+		if(!userId.equals(id.intValue())) {
+			throw new AccessDeniedException("User not authorized to access this resource");
+		}
 		userService.deleteUserById(id);
 		return "User deleted successfully.";
 	}
